pyproject.toml
[tool.setuptools]
include-package-data = true
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "config-saver"
version = "1.0.0"
description = "Python CLI tool for validating, converting, and managing YAML/JSON configs, with tar compression utilities"
authors = [
    { name = "amt911", email = "a.merlo.truji10@gmail.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pydantic",
    "colorama",
    "tqdm",
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/amt911/config-saver"

[project.scripts]
config-saver = "config_saver.__main__:main"

[tool.setuptools.packages.find]
where = ["config_saver"]

